Take initial world matrix from filepath
Add another argument that takes the path of the file
that stores the world matrix to start with. Flag is
"start-world-from-file" and should get the path of the
files that stores the map as an array in ascii characters.
It is assumed that the file conforms to the following rules:

1. It must contain a rectangular map of 2 distinct ascii
characters.
2. The character to represent a dead cell is '.'
3. The character to represent a live cell is 'M'
4. The file must end with a newline to provide the right
metric of number of lines in the map that will be stored
dynamically